+++ app/lib/components/blocks/entry_form.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import '../../model/form_model.dart';
 import '../../model/score_evaluation_message.pb.dart';
 import '../../view_model/providers.dart';
-import '../../view_model/form_state_notifier.dart' as form;
+import '../../view_model/form_view_model.dart' as FormViewModel;
 
 class EntryForm extends HookConsumerWidget {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
@@ -161,7 +161,7 @@ class EntryForm extends HookConsumerWidget {
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 16.0),
                 child: ((){
-                  if (_state is form.FormInitial){
+                  if (_state is FormViewModel.FormInitial){
                     return ElevatedButton(
                       onPressed: () {
                         if (!_formKey.currentState!.validate()) {// validate form data when button pressed
@@ -185,11 +185,11 @@ class EntryForm extends HookConsumerWidget {
                       },
                       child: const Text("Submit"),
                     );
-                  } else if (_state is form.FormSubmitting){
+                  } else if (_state is FormViewModel.FormSubmitting){
                     return CircularProgressIndicator();
-                  } else if (_state is form.FormSubmitted){
+                  } else if (_state is FormViewModel.FormSubmitted){
                     return Text("form successfully submitted");
-                  } else if (_state is form.FormError){
+                  } else if (_state is FormViewModel.FormError){
                     return Text(_state.message);
                   } else {
                   return Text("unexpected error occured\n${_state}");
@@ -199,7 +199,7 @@ class EntryForm extends HookConsumerWidget {
               Padding(
                 padding: const EdgeInsets.symmetric(vertical: 0.0),
                 child: ((){
-                  if (_state is form.FormError || _state is form.FormSubmitted){
+                  if (_state is FormViewModel.FormError || _state is FormViewModel.FormSubmitted){
                     return ElevatedButton(
                       onPressed: (){
                         ref.read(formStateNotifierProvider.notifier).initializeState();
